# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nanoni Engine",
    "description": "Nanoni control plane and local-media orchestration foundation",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/commerce/orders": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Order Create",
        "operationId": "order_create_api_v1_commerce_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/orders/{order_id}": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Order Get",
        "operationId": "order_get_api_v1_commerce_orders__order_id__get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/orders/{order_id}/payment": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Order Payment",
        "operationId": "order_payment_api_v1_commerce_orders__order_id__payment_get",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/mock-payments/{payment_id}/confirm": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Mock Confirm",
        "operationId": "mock_confirm_api_v1_commerce_mock_payments__payment_id__confirm_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmMockPayment"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/orders/{order_id}/fulfill": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Fulfill",
        "operationId": "fulfill_api_v1_commerce_orders__order_id__fulfill_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Order Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntitlementRead"
                  },
                  "title": "Response Fulfill Api V1 Commerce Orders  Order Id  Fulfill Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/webhooks/{provider_name}": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Payment Webhook",
        "operationId": "payment_webhook_api_v1_commerce_webhooks__provider_name__post",
        "parameters": [
          {
            "name": "provider_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Provider Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/payments/{payment_id}/recheck": {
      "post": {
        "tags": [
          "commerce"
        ],
        "summary": "Payment Recheck",
        "operationId": "payment_recheck_api_v1_commerce_payments__payment_id__recheck_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/commerce/entitlements": {
      "get": {
        "tags": [
          "commerce"
        ],
        "summary": "Entitlement List",
        "operationId": "entitlement_list_api_v1_commerce_entitlements_get",
        "parameters": [
          {
            "name": "customer_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Customer Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntitlementRead"
                  },
                  "title": "Response Entitlement List Api V1 Commerce Entitlements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmMockPayment": {
        "properties": {
          "provider_event_id": {
            "type": "string",
            "title": "Provider Event Id"
          }
        },
        "type": "object",
        "required": [
          "provider_event_id"
        ],
        "title": "ConfirmMockPayment"
      },
...
      "EntitlementRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "product_id": {
            "type": "string",
            "title": "Product Id"
          },
          "community_version_id": {
            "type": "string",
            "title": "Community Version Id"
          },
          "price_plan_id": {
            "type": "string",
            "title": "Price Plan Id"
          },
          "starts_at": {
            "type": "string",
            "format": "date-time",
            "title": "Starts At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "product_id",
          "community_version_id",
          "price_plan_id",
          "starts_at",
          "expires_at",
          "status"
        ],
        "title": "EntitlementRead"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderCreate": {
        "properties": {
          "telegram_user_id": {
            "type": "string",
            "title": "Telegram User Id"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "price_plan_id": {
            "type": "string",
            "title": "Price Plan Id"
          },
          "idempotency_key": {
            "type": "string",
            "maxLength": 120,
            "minLength": 8,
            "title": "Idempotency Key"
          }
        },
        "type": "object",
        "required": [
          "telegram_user_id",
          "price_plan_id",
          "idempotency_key"
        ],
        "title": "OrderCreate"
      },
      "OrderRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "customer_id": {
            "type": "string",
            "title": "Customer Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "total_amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Amount"
          },
          "idempotency_key": {
            "type": "string",
            "title": "Idempotency Key"
          },
          "checkout_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkout Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "customer_id",
          "status",
          "currency",
          "total_amount",
          "idempotency_key",
          "checkout_url"
        ],
        "title": "OrderRead"
      },
      "PaymentRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "title": "Order Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "provider_charge_id": {
            "type": "string",
            "title": "Provider Charge Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "checkout_payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Checkout Payload"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "provider",
          "provider_charge_id",
          "status",
          "amount",
          "currency",
          "checkout_payload",
          "confirmed_at"
        ],
        "title": "PaymentRead"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}